pyproject.toml
[project]
name = "beckings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.43.0",
    "cookiecutter>=2.6.0",
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-cotton>=2.0.3",
    "django-htmx>=1.23.0",
    "django-redis>=5.4.0",
    "django-tailwind>=4.0.1",
    "django-template-partials>=24.4",
    "djangorestframework>=3.16.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg2[binary]>=2.9.10",
    "psycopg[binary]>=3.2.7",
    "python-decouple>=3.8",
    "rav>=0.0.9",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar>=5.1.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "time-machine>=2.16.0",
]

[tool.pytest.init_options]
DJANGO_SETTINGS_MODULE = "src.beckings.settings"
python_files = [ "src/tests/*", "src/tests/test_*.py" ]
